Add MetricsManager and refactor controller to use manager
